requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
requests==2.33.0
numpy==2.4.3
pandas==3.0.1
scikit-learn==1.8.0
joblib==1.5.3
pydantic==2.12.5
langchain==1.2.13
langchain-ollama==1.0.1
langchain-community==0.4.1
langgraph==1.1.3
faiss-cpu==1.13.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
pytest==9.0.2
httpx==0.28.1
ruff
